Добавляет комментарии в модальное окно просмотра.
